chore(sourcemap): convert inputSourceMap to plain js object
@babel/core recently made a change where the sourcemap is no longer a plain js
object, but instead is a class SourceMap. This breaks when the coverage data
gets passed into valueToNode by istanbul-lib-instrument because the coverage
data is no longer entirely plain json objects (this.cov.toJSON() just returns
the data and doesn't do any conversion).

In this PR we are checking if the input sourcemap is a plain js object, and
converting it if it is not.